Split Bootstraping for Page and Event into multiple templates
