+++ src/Tribe/Views/V2/Template/Event.php
@@ -1,0 +1,43 @@
+<?php
+/**
+ * Initializer for The Events Calendar for the template structure using Event
+ *
+ * Can be changed on Events > Settings > Display
+ *
+ * @since   TBD
+ *
+ * @package Tribe\Events\Views\V2
+ */
+namespace Tribe\Events\Views\V2\Template;
+
+use Tribe__Events__Main as TEC;
+use Tribe\Events\Views\V2\Index as Index;
+
+class Event {
+
+	/**
+	 * Hook the Template Hijack into WordPress methods
+	 *
+	 * @since  TBD
+	 *
+	 * @return void
+	 */
+	public function hook() {
+
+	}
+
+	/**
+	 * Determines the Path for the PHP file to be used as the main template
+	 * For Page base template setting it will select from theme or child theme
+	 *
+	 * @todo  Integrate with Template + Context classes
+	 *
+	 * @since  TBD
+	 *
+	 * @return string
+	 */
+	public function get_path() {
+		return ( new Index() )->get_template_file();
+	}
+
+}
